# Optimize UUID to/from string conversion