### Benchmark*

https://gist.github.com/9il/7480f54792e22b1e916f

#### LDC 

based on DMD v2.067.1 and LLVM 3.6.0
```
$ldmd2 -O -inline -release -boundscheck=off -run uuid_bench.d
Old constructor: 397 μs and 3 hnsecs
New constructor: 41 μs and 9 hnsecs
Old toString: 47 μs and 3 hnsecs
New toString: 21 μs and 8 hnsecs
```

#### DMD

2.068.2
```
$dmd -O -inline -release -boundscheck=off -run uuid_bench.d
Old constructor: 1 ms, 852 μs, and 7 hnsecs
New constructor: 75 μs and 4 hnsecs
Old toString: 184 μs and 7 hnsecs
New toString: 51 μs and 6 hnsecs
```
_______
*time for 1000 conversions.