@9il 2015-10-02
ping @DmitryOlshansky: second part of UUID optimisation.