@DmitryOlshansky 2015-10-03
Why can't it use standard toString with delegate?