@9il 2015-10-03
It has 3 `toString` prototypes:

1. First
  1. specialised
  2. for `@safe pure nothrow @nogc` code
  3. pure zero-copy code
2.  Second 
  1. for `std.format` functionality.
  2. standart converting without allocations
3. Third
  1. for `@safe pure nothrow` code
  2. simple API

Concept like that was used before this PR, but [analog of the first prototype](https://github.com/D-Programming-Language/phobos/pull/3686/files#diff-6402ed99078abad75d161c868510b688L141) was private. It is useful for optimised server applications and DB bindings.

Third is simplest and for `@safe pure nothrow` code.
```D
/**
 * Return the UUID as a string in the canonical form.
 */
@safe pure nothrow @nogc void toString(ref char[36] result) const
{ ... }
///ditto
void toString(scope void delegate(const(char)[]) sink) const
{
    sink(_toString());
    char[36] result = void;
    toString(result);
    sink(result[]);
}
///ditto
@safe pure nothrow string toString() const
{
    return _toString().idup;
    char[36] result = void;
    toString(result);
    return result[].idup;
}
```