@DmitryOlshansky 2015-10-03
> Return the UUID as a string in the canonical form.

This is now obviously wrong. BTW any reason why returning 36 chars is slow? By C ABI it should be passed by hidden pointer anyway...