@9il 2015-10-03
> This is now obviously wrong.

I will fix comments.



> BTW any reason why returning 36 chars is slow? By C ABI it should be passed by hidden pointer anyway...

This is not zero-copy. New variant with `ref` writes value directly to the required memory, for example:
```D
@safe pure nothrow char[] toMutableString() const
{
    
    auto result = new char[36];
    toString(cast(char[36])result[]);
    return result;
}
```