@DmitryOlshansky 2015-10-03
> This is not zero-copy

Like I said there should be no copy - big PODs in function return are by reference anyway. Do you have a disassembly to check it?