@9il 2015-10-03
> Like I said there should be no copy - big PODs in function return are by reference anyway. Do you have a disassembly to check it?

Zero-copy not for this function, but for a code that use this function:

```D
///New, zero-copy
@safe pure nothrow char[] toMutableString() const
{
    auto result = new char[36];
    toString(cast(char[36])result[]); //writes directly to GC allocated memory.
    return result;
}
///Old
@safe pure nothrow char[] toMutableString() const
{
    char[36] local = _toString; //OK, inlined
    return local.dup;            //Copied (!) to GC allocated memory.
}
```