@9il 2015-10-03
> >  Return the UUID as a string in the canonical form.

> This is now obviously wrong. BTW any reason why returning 36 chars is slow? By C ABI it should be passed by hidden pointer anyway...

Fixed https://github.com/9il/phobos/commit/f385510c0f2a6092d2af81edf35a770f501779ff