@DmitryOlshansky 2015-10-03
> Zero-copy not for this function, but for a code that use this function:

Every time I see `cast(char[X])` I cringe it's not any safer then just `assert(slice.length >= 36)` and certainly not easier to use.