@9il 2015-10-03
> Every time I see cast(char[X]) I cringe. It's not any safer then just assert(slice.length >= 36) and certainly not easier to use.

I agree. Should the first function be undocumented?