@DmitryOlshansky 2015-10-03
> I agree. Should the first function be undocumented?

No just let it accept normal slice or rather an output range (a slice is output range)?