@DmitryOlshansky 2015-10-03
Undocumented things do get used and then we change them and we have a regerssion. Please no undocumented stuff for "those in the know"