@9il 2015-10-03
> No just let it accept normal slice or rather an output range (a slice is output range)?

Prototype for slice: https://github.com/9il/phobos/commit/dbdc6ef7c74a22677a7655417fb449e9f83b9539#diff-6402ed99078abad75d161c868510b688R863. Output ranges can not be used because loop unrolling.