@DmitryOlshansky 2015-10-03
> Output ranges can not be used because loop unrolling.

Just use specialization?