@9il 2015-10-03
> Just use specialization?

Done 318cb998e1cb547a2fac4ff365d8387097e1e142