@DmitryOlshansky 2015-10-03
> Done 318cb99

Cool and if there would be tests that test both (w|d)char[] and something else I'm ready to pull this ;)