@9il 2015-10-03
> if there would be tests that test both (w|d)char[] and something else

I have update constraints, so `AsciiChar` can be used. 58982e4e1cb9c342809c26f0cd6ab59d36a2e3b5