@DmitryOlshansky 2015-10-03
Auto-merge toggled on